+++ frontend/src/features/notes/components/NoteCreate/index.tsx
@@ -23,7 +23,7 @@ import type { Note } from '@/features/notes/types';
 import { refreshImageUrls, uploadImage } from '@/lib/api/file';
 import FileHandler from '@tiptap-pro/extension-file-handler';
 import Image from '@tiptap/extension-image';
-import { EditorContent, useEditor } from '@tiptap/react';
+import { Editor, EditorContent, useEditor } from '@tiptap/react';
 import StarterKit from '@tiptap/starter-kit';
 import { ArrowLeft, Save } from 'lucide-react';
 import React, { useEffect, useState, useRef } from 'react';
@@ -39,6 +39,16 @@ export type NoteCreateProps = {
   projectId: string;
   projectUrlId?: string; // プロジェクトのURL ID
   noteId?: string; // 編集時のノートID
+  handleSubmit: (e: React.FormEvent) => void;
+  handleCancel: () => void;
+  title: string;
+  setTitle: (title: string) => void;
+  content: string;
+  setContent: (content: string) => void;
+  debouncedContent: string;
+  isUploading: boolean;
+  editor: Editor | null;
+  isRefreshingImages: boolean;
 };
 
 const NoteCreate: React.FC<NoteCreateProps> = ({
@@ -51,189 +61,25 @@ const NoteCreate: React.FC<NoteCreateProps> = ({
   projectId,
   projectUrlId,
   noteId,
+  handleSubmit,
+  handleCancel,
+  title,
+  setTitle,
+  content,
+  setContent,
+  debouncedContent,
+  isUploading,
+  editor,
+  isRefreshingImages
 }) => {
-  const [title, setTitle] = React.useState(initialTitle);
-  // tiptap表示用content
-  const [content, setContent] = React.useState('');
-  // 即座にcontentの内容に対して処理を行うと負荷がかかるため、500ms後にcontentと同期させてdebouncedContentの値でタグ抽出等の処理を行っている
-  const [debouncedContent, setDebouncedContent] = useState('');
-  const debounceTimerRef = useRef<NodeJS.Timeout | null>(null);
-  const [isUploading, setIsUploading] = useState(false);
-  const [editorInitialized, setEditorInitialized] = useState(false);
-  const [isRefreshingImages, setIsRefreshingImages] = useState(false);
-
-  // 初期コンテンツの画像URLを更新
-  useEffect(() => {
-    const updateInitialContent = async () => {
-      if (initialContent) {
-        setIsRefreshingImages(true);
-        try {
-          // 初期コンテンツ内の画像URLを更新
-          const updatedContent = await refreshImageUrls(initialContent);
-          setContent(updatedContent);
-          setDebouncedContent(updatedContent);
-        } catch (error) {
-          console.error('画像URL更新エラー:', error);
-          setContent(initialContent);
-          setDebouncedContent(initialContent);
-        } finally {
-          setIsRefreshingImages(false);
-        }
-      } else {
-        setContent('');
+      // エディタが初期化されていない場合はローディング表示
+      if (!editor && !isRefreshingImages) {
+        return (
+          <div className="flex justify-center items-center h-64">
+            <div className="animate-spin rounded-full h-8 w-8 border-t-2 border-b-2 border-primary" />
+          </div>
+        );
       }
-    };
-
-    updateInitialContent();
-  }, [initialContent]);
-
-  const editor = useEditor({
-    extensions: [
-      StarterKit,
-      Image.configure({
-        allowBase64: true, // 一時的な画像表示のために残す
-        inline: false,
-      }),
-      Tag, // タグ拡張を追加
-      FileHandler.configure({
-        allowedMimeTypes: [
-          'image/png',
-          'image/jpeg',
-          'image/gif',
-          'image/webp',
-          'image/svg+xml',
-        ],
-        onDrop: async (currentEditor, files, pos) => {
-          if (!files.length) return;
-
-          setIsUploading(true);
-
-          try {
-            for (const file of files) {
-              if (file.size > 10 * 1024 * 1024) {
-                // 10MB制限
-                alert('画像サイズは10MB以下にしてください');
-                continue;
-              }
-
-              // Supabase Storageに画像をアップロード
-              const imageUrl = await uploadImage(projectId, file);
-
-              // エディタに画像を挿入
-              currentEditor
-                .chain()
-                .insertContentAt(pos, {
-                  type: 'image',
-                  attrs: {
-                    src: imageUrl,
-                    alt: file.name,
-                  },
-                })
-                .focus()
-                .run();
-            }
-          } catch (error) {
-            console.error('画像アップロードエラー:', error);
-            alert('画像のアップロードに失敗しました');
-          } finally {
-            setIsUploading(false);
-          }
-        },
-        onPaste: async (currentEditor, files, htmlContent) => {
-          if (!files.length) return;
-
-          setIsUploading(true);
-
-          try {
-            for (const file of files) {
-              if (file.size > 10 * 1024 * 1024) {
-                // 10MB制限
-                alert('画像サイズは10MB以下にしてください');
-                continue;
-              }
-
-              // Supabase Storageに画像をアップロード
-              const imageUrl = await uploadImage(projectId, file);
-
-              // エディタに画像を挿入
-              currentEditor
-                .chain()
-                .insertContent({
-                  type: 'image',
-                  attrs: {
-                    src: imageUrl,
-                    alt: file.name,
-                  },
-                })
-                .focus()
-                .run();
-            }
-          } catch (error) {
-            console.error('画像アップロードエラー:', error);
-            alert('画像のアップロードに失敗しました');
-          } finally {
-            setIsUploading(false);
-          }
-        },
-      }),
-    ],
-    content: content,
-    onUpdate: ({ editor }) => {
-      const html = editor.getHTML();
-      setContent(html);
-
-      // コンテンツの変更を遅延させて関連ノートの検索を最適化
-      if (debounceTimerRef.current) {
-        clearTimeout(debounceTimerRef.current);
-      }
-
-      debounceTimerRef.current = setTimeout(() => {
-        setDebouncedContent(html);
-      }, 500); // 500ms後に更新
-    },
-    editorProps: {
-      attributes: {
-        class: 'prose focus:outline-none max-w-none',
-      },
-    },
-    // SSR対応のため、即時レンダリングをオフに
-    immediatelyRender: false,
-  });
-
-  // エディタが初期化されたときに初期コンテンツを設定
-  useEffect(() => {
-    if (editor && content && !editorInitialized) {
-      editor.commands.setContent(content);
-      setEditorInitialized(true);
-    }
-  }, [editor, content, editorInitialized]);
-
-  // コンポーネントがアンマウントされるときにタイマーをクリア
-  useEffect(() => {
-    return () => {
-      if (debounceTimerRef.current) {
-        clearTimeout(debounceTimerRef.current);
-      }
-    };
-  }, []);
-
-  // エディタが初期化されていない場合はローディング表示
-  if (!editor && !isRefreshingImages) {
-    return (
-      <div className="flex justify-center items-center h-64">
-        <div className="animate-spin rounded-full h-8 w-8 border-t-2 border-b-2 border-primary" />
-      </div>
-    );
-  }
-
-  // フォーム送信処理
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    onSubmit({
-      title,
-      content,
-    });
-  };
 
   return (
     <Card className="w-full max-w-4xl mx-auto">
